move version dumping implementation to own internal files
