+++ src/setuptools_scm/__init__.py
@@ -1,26 +1,20 @@
 """
-:copyright: 2010-2015 by Ronny Pfannschmidt
+:copyright: 2010-2023 by Ronny Pfannschmidt
 :license: MIT
 """
 from __future__ import annotations
 
-import os
 import re
 import warnings
 from typing import Any
 from typing import Pattern
 from typing import TYPE_CHECKING
 
+from . import _config
 from ._config import Configuration
-from ._config import DEFAULT_LOCAL_SCHEME
-from ._config import DEFAULT_TAG_REGEX
-from ._config import DEFAULT_VERSION_SCHEME
 from ._entrypoints import _version_from_entrypoints
 from ._overrides import _read_pretended_version_for
-from ._overrides import PRETEND_KEY
-from ._overrides import PRETEND_KEY_NAMED
 from ._version_cls import _validate_version_cls
-from ._version_cls import _version_as_tuple
 from ._version_cls import NonNormalizedVersion
 from ._version_cls import Version
 from .version import format_version as _format_version
@@ -31,17 +25,6 @@ if TYPE_CHECKING:
     from . import _types as _t
 
 
-TEMPLATES = {
-    ".py": """\
-# file generated by setuptools_scm
-# don't change, don't track in version control
-__version__ = version = {version!r}
-__version_tuple__ = version_tuple = {version_tuple!r}
-""",
-    ".txt": "{version}",
-}
-
-
 def dump_version(
     root: _t.PathT,
     version: str,
@@ -49,33 +32,9 @@ def dump_version(
     template: str | None = None,
     scm_version: ScmVersion | None = None,
 ) -> None:
-    assert isinstance(version, str)
-    target = os.path.normpath(os.path.join(root, write_to))
-    ext = os.path.splitext(target)[1]
-    template = template or TEMPLATES.get(ext)
-    from ._log import log
+    from ._integration.dump_version import dump_version as real
 
-    log.debug("dump %s into %s", version, write_to)
-    if template is None:
-        raise ValueError(
-            "bad file format: '{}' (of {}) \nonly *.txt and *.py are supported".format(
-                os.path.splitext(target)[1], target
-            )
-        )
-    version_tuple = _version_as_tuple(version)
-
-    if scm_version is not None:
-        with open(target, "w") as fp:
-            fp.write(
-                template.format(
-                    version=version,
-                    version_tuple=version_tuple,
-                    scm_version=scm_version,
-                )
-            )
-    else:
-        with open(target, "w") as fp:
-            fp.write(template.format(version=version, version_tuple=version_tuple))
+    return real(root, version, write_to, template, scm_version)
 
 
 def _do_parse(config: Configuration) -> _t.SCMVERSION | None:
@@ -121,12 +80,12 @@ def _version_missing(config: Configuration) -> NoReturn:
 
 def get_version(
     root: _t.PathT = ".",
-    version_scheme: _t.VERSION_SCHEME = DEFAULT_VERSION_SCHEME,
-    local_scheme: _t.VERSION_SCHEME = DEFAULT_LOCAL_SCHEME,
+    version_scheme: _t.VERSION_SCHEME = _config.DEFAULT_VERSION_SCHEME,
+    local_scheme: _t.VERSION_SCHEME = _config.DEFAULT_LOCAL_SCHEME,
     write_to: _t.PathT | None = None,
     write_to_template: str | None = None,
     relative_to: str | None = None,
-    tag_regex: str | Pattern[str] = DEFAULT_TAG_REGEX,
+    tag_regex: str | Pattern[str] = _config.DEFAULT_TAG_REGEX,
     parentdir_prefix_version: str | None = None,
     fallback_version: str | None = None,
     fallback_root: _t.PathT = ".",
@@ -152,7 +111,7 @@ def get_version(
                     "empty regex for tag regex is invalid, using default"
                 )
             )
-            tag_regex = DEFAULT_TAG_REGEX
+            tag_regex = _config.DEFAULT_TAG_REGEX
         else:
             tag_regex = re.compile(tag_regex)
     config = Configuration(**locals())
@@ -189,11 +148,6 @@ __all__ = [
     "get_version",
     "dump_version",
     "Configuration",
-    "DEFAULT_VERSION_SCHEME",
-    "DEFAULT_LOCAL_SCHEME",
-    "DEFAULT_TAG_REGEX",
-    "PRETEND_KEY",
-    "PRETEND_KEY_NAMED",
     "Version",
     "NonNormalizedVersion",
 ]
