+++ testing/test_basic_api.py
@@ -143,7 +143,7 @@ def test_get_version_blank_tag_regex() -> None:
     "version", ["1.0", "1.2.3.dev1+ge871260", "1.2.3.dev15+ge871260.d20180625", "2345"]
 )
 def test_pretended(version: str, monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setenv(setuptools_scm.PRETEND_KEY, version)
+    monkeypatch.setenv(setuptools_scm._overrides.PRETEND_KEY, version)
     assert setuptools_scm.get_version() == version
