+++ testing/test_functions.py
@@ -7,7 +7,7 @@ import pytest
 from setuptools_scm import Configuration
 from setuptools_scm import dump_version
 from setuptools_scm import get_version
-from setuptools_scm import PRETEND_KEY
+from setuptools_scm._overrides import PRETEND_KEY
 from setuptools_scm._run_cmd import has_command
 from setuptools_scm.version import format_version
 from setuptools_scm.version import guess_next_version
